backend/requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[argon2]
argon2-cffi
sqlalchemy                  
pymysql                    
python-dotenv 
pydantic-settings             
email-validator